refactor(types): extract mod tests to sibling file
